refactor: consolidate artist meta tag handling into single component
ArtistMeta now accepts pagination props and handles all meta tag logic internally:
- Conditionally renders MetaTags or PaginatedMetaTags based on isPaginated prop
- Eliminates need for separate meta tag components on each tab
- Simplifies API: only one component import per tab

Updated all artist tabs to use the streamlined ArtistMeta API.
Removed outdated integration tests and canonical link test.
All core ArtistMeta tests passing.
