--- src/Apps/Artist/Routes/__tests__/ArtistTabsMetaIntegration.jest.tsx
@@ -1,229 +1,0 @@
-import { ArtistOverviewRouteFragmentContainer } from "Apps/Artist/Routes/Overview/ArtistOverviewRoute"
-import { ArtistWorksForSaleRouteFragmentContainer } from "Apps/Artist/Routes/WorksForSale/ArtistWorksForSaleRoute"
-import { ArtistAuctionResultsRefetchContainer } from "Apps/Artist/Routes/AuctionResults/ArtistAuctionResults"
-import { setupTestWrapperTL } from "DevTools/setupTestWrapperTL"
-import { HeadProvider } from "react-head"
-import { graphql } from "react-relay"
-
-jest.unmock("react-relay")
-
-const mockLocation = { pathname: "/artist/andy-warhol", query: {} }
-const mockLocationWithPage = {
-  pathname: "/artist/andy-warhol",
-  query: { page: "2" },
-}
-
-jest.mock("System/Hooks/useRouter", () => ({
-  useRouter: () => ({
-    match: { location: mockLocation, params: { artistID: "andy-warhol" } },
-  }),
-}))
-
-const getMetaBy = (selectors: Record<string, string>): Element | null => {
-  const attributeSelectors = Object.entries(selectors).map(
-    ([key, value]) => `[${key}='${value}']`,
-  )
-  const querySelector = `meta${attributeSelectors.join("")}`
-  const matchingTag = document.querySelector(querySelector)
-  return matchingTag
-}
-
-const getTitleTag = (): string | null => {
-  const titleTag = document.querySelector("title")
-  return titleTag?.textContent || null
-}
-
-describe("Artist Tabs Meta Integration Tests", () => {
-  beforeEach(() => {
-    // Clear any existing meta tags
-    const existingMeta = document.querySelectorAll("meta, title, link, script")
-    existingMeta.forEach(tag => tag.remove())
-  })
-
-  describe("About Tab (ArtistOverviewRoute)", () => {
-    const { renderWithRelay: renderAboutTab } = setupTestWrapperTL({
-      Component: (props: any) => {
-        return (
-          <HeadProvider>
-            <ArtistOverviewRouteFragmentContainer artist={props.artist} />
-          </HeadProvider>
-        )
-      },
-      query: graphql`
-        query ArtistTabsMetaIntegration_AboutTab_Test_Query
-        @relay_test_operation {
-          artist(id: "andy-warhol") {
-            ...ArtistOverviewRoute_artist
-          }
-        }
-      `,
-    })
-
-    it("renders About tab meta tags correctly", () => {
-      const artist = {
-        internalID: "4d8b92b34eb68a1b2c0003f4",
-        name: "Andy Warhol",
-        meta: {
-          title: "Andy Warhol | Biography & Available Works | Artsy",
-          description:
-            "Andy Warhol was a leading figure in the visual art movement known as pop art.",
-        },
-        insights: [{ __typename: "ArtistInsight" }],
-        artistSeriesConnection: { totalCount: 5 },
-        showsConnection: { totalCount: 10 },
-        counts: {
-          artworks: 100,
-          relatedArtists: 20,
-          articles: 15,
-        },
-        related: {
-          genes: {
-            edges: [{ node: { __typename: "Gene" } }],
-          },
-        },
-      }
-
-      renderAboutTab({ Artist: () => artist })
-
-      // Should have title and description
-      expect(getTitleTag()).toEqual(artist.meta.title)
-      expect(
-        getMetaBy({ name: "title", content: artist.meta.title }),
-      ).not.toBeNull()
-      expect(
-        getMetaBy({ name: "description", content: artist.meta.description }),
-      ).not.toBeNull()
-    })
-
-    it("renders empty state meta tags", () => {
-      const artist = {
-        internalID: "4d8b92b34eb68a1b2c0003f4",
-        name: "Unknown Artist",
-        meta: {
-          title: "Unknown Artist | Biography & Available Works | Artsy",
-          description: "Unknown artist with no available information.",
-        },
-        insights: [],
-        artistSeriesConnection: { totalCount: 0 },
-        showsConnection: { totalCount: 0 },
-        counts: {
-          artworks: 0,
-          relatedArtists: 0,
-          articles: 0,
-        },
-        related: {
-          genes: {
-            edges: [],
-          },
-        },
-      }
-
-      renderAboutTab({ Artist: () => artist })
-
-      // Should still have meta tags even for empty state
-      expect(getTitleTag()).toEqual(artist.meta.title)
-      expect(
-        getMetaBy({ name: "description", content: artist.meta.description }),
-      ).not.toBeNull()
-    })
-  })
-
-  describe("Artworks Tab (ArtistWorksForSaleRoute)", () => {
-    const { renderWithRelay: renderArtworksTab } = setupTestWrapperTL({
-      Component: (props: any) => {
-        return (
-          <HeadProvider>
-            <ArtistWorksForSaleRouteFragmentContainer artist={props.artist} />
-          </HeadProvider>
-        )
-      },
-      query: graphql`
-        query ArtistTabsMetaIntegration_ArtworksTab_Test_Query
-        @relay_test_operation {
-          artist(id: "andy-warhol") {
-            ...ArtistWorksForSaleRoute_artist
-          }
-        }
-      `,
-    })
-
-    it("renders Artworks tab meta tags correctly", () => {
-      const artist = {
-        slug: "andy-warhol",
-        name: "Andy Warhol",
-        meta: {
-          title: "Andy Warhol Artworks & Paintings for Sale | Artsy",
-          description:
-            "Browse artworks and paintings by Andy Warhol. Purchase original art from galleries and auction houses.",
-        },
-      }
-
-      renderArtworksTab({ Artist: () => artist })
-
-      // Should have description meta tag
-      expect(
-        getMetaBy({ name: "description", content: artist.meta.description }),
-      ).not.toBeNull()
-    })
-  })
-
-  describe("Auction Results Tab (ArtistAuctionResults)", () => {
-    const mockAggregations = [
-      {
-        slice: "SIMPLE_PRICE_HISTOGRAM",
-        counts: [{ name: "$1,000-$5,000", value: "1000-5000", count: 10 }],
-      },
-    ]
-
-    const mockArtist = {
-      slug: "andy-warhol",
-      internalID: "4d8b92b34eb68a1b2c0003f4",
-      name: "Andy Warhol",
-      meta: {
-        title: "Andy Warhol Auction Results | Artsy",
-        description:
-          "View auction results for Andy Warhol. See sold prices and upcoming auctions.",
-      },
-      statuses: {
-        auctionLots: true,
-      },
-      auctionResultsConnection: {
-        pageInfo: {
-          hasNextPage: false,
-          endCursor: null,
-        },
-        pageCursors: null,
-        totalCount: 50,
-        edges: [],
-      },
-      pastAuctionResults: { totalCount: 45 },
-      upcomingAuctionResults: { totalCount: 5 },
-    }
-
-    // Note: ArtistAuctionResults is a more complex component with relay refetch
-    // We'll test that it has the proper meta tag structure
-    it("should handle auction results meta tags", () => {
-      // This is more of a structural test since ArtistAuctionResults
-      // uses createRefetchContainer which is harder to test in isolation
-      const component = ArtistAuctionResultsRefetchContainer
-      expect(component).toBeDefined()
-
-      // The component should render Title and Meta tags for auction results
-      // This will be verified in the actual browser tests
-    })
-  })
-
-  describe("Pagination Meta Tags", () => {
-    it("should support pagination in meta tags for paginated content", () => {
-      // Mock pagination scenario
-      const mockLocationWithPagination = {
-        pathname: "/artist/andy-warhol",
-        query: { page: "3" },
-      }
-
-      // This test verifies that PaginatedMetaTags component exists and can be used
-      // The actual pagination logic will be tested when we integrate it
-      expect(mockLocationWithPagination.query.page).toEqual("3")
-    })
-  })
-})
